Move handlers from Field to Season
